refactor: Move executor definition to the ref package
